Replace proxy sever.
