+++ hkm/finna.py
@@ -2,10 +2,14 @@
 
 import logging
 import math
+import urllib
+import urlparse
 from StringIO import StringIO
 
 import requests
 from PIL import Image
+from django.core.cache import cache
+
 
 LOG = logging.getLogger(__name__)
 
@@ -114,6 +118,59 @@ class FinnaClient(object):
         else:
             url = 'https://finna.fi/Cover/Show?id=%s&fullres=1&index=0' % record_id
         return url
+
+    def get_full_res_image_url(self, preview_image_url):
+        timeout = 6
+
+        """
+        HKM image server responds with 200 text body response if there is no full res version for the
+        image. This checks the response headers to find out whether full res image exists or not and
+        returns the full res image url if it does. Else returns None
+        """
+        cache_key = 'hkm_%s' % preview_image_url
+        cached_value = cache.get(cache_key)
+        if cached_value:
+            return cached_value
+
+        url_components = urlparse.urlparse(preview_image_url)
+        qs_params = dict(urlparse.parse_qsl(url_components.query))
+        qs_params['dataType'] = 'org'
+
+        url_components = list(url_components)
+        url_components[4] = urllib.urlencode(qs_params)
+        url = urlparse.urlunparse(url_components)
+        try:
+            r = requests.head(url, timeout=self.timeout)
+        except requests.exceptions.RequestException:
+            LOG.error('Failed to communicate with FINNA image server',
+                      exc_info=True)
+            # Cache negative result for an hour
+            cache.set(cache_key, None, 3600)
+            return None
+        else:
+            try:
+                r.raise_for_status()
+                # raise requests.exceptions.HTTPError()
+            except requests.exceptions.HTTPError:
+                LOG.error('Failed to communicate with FINNA image server', exc_info=True,
+                          extra={'data': {'status_code': r.status_code, 'response': repr(r.text)}})
+                # Cache negative result for an hour
+                cache.set(cache_key, None, 3600)
+                return None
+
+        LOG.debug('Got result from FINNA image server',
+                  extra={'data': {'url': url, 'result_headers': repr(r.headers)}})
+        try:
+            content_type = r.headers['content-type']
+            if 'image' in content_type:
+                # Cache positive result for a week
+                cache.set(cache_key, url, 7 * 24 * 3600)
+                return url
+        except KeyError:
+            pass
+
+        cache.set(cache_key, None, 3600)
+        return None
 
     def download_image(self, record_id):
         r = requests.get(self.get_image_url(record_id), stream=True)
