+++ hkm/models/models.py
@@ -27,7 +27,6 @@ from parler.models import TranslatableModel, TranslatedFields
 from phonenumber_field.modelfields import PhoneNumberField
 
 from hkm.finna import DEFAULT_CLIENT as FINNA
-from hkm.hkm_client import DEFAULT_CLIENT as HKM
 from hkm.models.campaigns import Campaign, CampaignStatus, CampaignCode, CodeUsage
 from hkm.paybyway_client import client as PBW
 from hkm.printmotor_client import client as PRINTMOTOR
@@ -213,7 +212,7 @@ class Record(OrderedModel, BaseModel):
             cache_key = '%s-record-preview-url' % record_url
             full_res_url = DEFAULT_CACHE.get(cache_key, None)
             if full_res_url is None:
-                full_res_url = HKM.get_full_res_image_url(
+                full_res_url = FINNA.get_full_res_image_url(
                     record_data['rawData']['thumbnail'])
                 DEFAULT_CACHE.set(cache_key, full_res_url, 60 * 15)
             else:
